requirements.txt
pyserial==3.4
opencv-python==4.4.0.42
numpy==1.19.1